+++ src/main/java/de/samply/tasks/PseudonymGeneratorDelegate.java
@@ -16,11 +16,13 @@ import de.samply.file.csv.updater.CsvUpdaterFactoryImpl;
 import de.samply.file.csv.updater.PivotedCsvRecordHeaderValues;
 import de.samply.pseudonymisation.Patient;
 import de.samply.pseudonymisation.PatientFields;
+import de.samply.pseudonymisation.PatientFieldsUtils;
 import de.samply.pseudonymisation.PseudonymisationClient;
 import de.samply.spring.MtbaConst;
 import de.samply.utils.PathsBundleUtils;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -45,81 +47,84 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
   private final Logger logger = LoggerFactory.getLogger(PseudonymGeneratorDelegate.class);
 
   private PseudonymisationClient pseudonymisationClient;
+  private PatientFieldsUtils patientFieldsUtils;
   private final CsvUpdaterFactory csvUpdaterFactory = new CsvUpdaterFactoryImpl();
-  private final String localIdCsvFilename;
-  private final String localIdCsvLocalId;
-  private final String idatCsvFilename;
+  private final String identityCsvLocalId;
+  private final String identityCsvFilename;
   private final String idManagerPseudonymIdType;
   private final Boolean removeIdat;
-  private final static String IDAT_CSV_TEMP_HEADER = "TEMP_PAT_ID";
+  private final static String IDENTITY_CSV_TEMP_HEADER = "TEMP_PAT_ID";
   private List<String> idatHeaders = new ArrayList<>();
   private Map<String, BiConsumer<PatientFields, String>> headerPatientFieldsMap = new HashMap<>();
 
 
   public PseudonymGeneratorDelegate(
-      @Value(MtbaConst.LOCAL_ID_CSV_FILENAME_SV) String localIdCsvFilename,
-      @Value(MtbaConst.LOCAL_ID_CSV_LOCAL_ID_HEADER_SV) String localIdCsvLocalId,
-      @Value(MtbaConst.IDAT_CSV_FILENAME_SV) String idatCsvFilename,
-      @Value(MtbaConst.IDAT_CSV_FIRST_NAME_HEADER_SV) String idatCsvFirstNameHeader,
-      @Value(MtbaConst.IDAT_CSV_LAST_NAME_HEADER_SV) String idatCsvLastNameHeader,
-      @Value(MtbaConst.IDAT_CSV_PREVIOUS_NAMES_HEADER_SV) String idatCsvPreviousNamesHeader,
-      @Value(MtbaConst.IDAT_CSV_BIRTHDAY_HEADER_SV) String idatCsvBirthdayHeader,
-      @Value(MtbaConst.IDAT_CSV_CITIZENSHIP_HEADER_SV) String idatCsvCitizenshipHeader,
-      @Value(MtbaConst.IDAT_CSV_GENDER_HEADER_SV) String idatCsvGenderHeader,
+      @Value(MtbaConst.IDENTITY_CSV_LOCAL_ID_HEADER_SV) String identityCsvLocalId,
+      @Value(MtbaConst.IDENTITY_CSV_FILENAME_SV) String identityCsvFilename,
+      @Value(MtbaConst.IDENTITY_CSV_FIRST_NAME_HEADER_SV) String identityCsvFirstNameHeader,
+      @Value(MtbaConst.IDENTITY_CSV_LAST_NAME_HEADER_SV) String identityCsvLastNameHeader,
+      @Value(MtbaConst.IDENTITY_CSV_PREVIOUS_NAMES_HEADER_SV) String identityCsvPreviousNamesHeader,
+      @Value(MtbaConst.IDENTITY_CSV_BIRTHDAY_HEADER_SV) String identityCsvBirthdayHeader,
+      @Value(MtbaConst.IDENTITY_CSV_CITIZENSHIP_HEADER_SV) String identityCsvCitizenshipHeader,
+      @Value(MtbaConst.IDENTITY_CSV_GENDER_HEADER_SV) String identityCsvGenderHeader,
       @Value(MtbaConst.ID_MANAGER_PSEUDONYM_ID_TYPE_SV) String idManagerPseudonymIdType,
       @Value(MtbaConst.CSV_REMOVE_IDAT_SV) String removeIdat,
-      @Autowired PseudonymisationClient pseudonymisationClient) {
+      @Autowired PseudonymisationClient pseudonymisationClient,
+      @Autowired PatientFieldsUtils patientFieldsUtils) {
 
     this.pseudonymisationClient = pseudonymisationClient;
+    this.patientFieldsUtils = patientFieldsUtils;
     this.removeIdat = Boolean.valueOf(removeIdat);
-    this.localIdCsvFilename = localIdCsvFilename;
-    this.localIdCsvLocalId = localIdCsvLocalId;
-    this.idatCsvFilename = idatCsvFilename;
+    this.identityCsvLocalId = identityCsvLocalId;
+    this.identityCsvFilename = identityCsvFilename;
     this.idManagerPseudonymIdType = idManagerPseudonymIdType;
-    String[] tempIdatHeaders = {idatCsvFirstNameHeader, idatCsvLastNameHeader,
-        idatCsvPreviousNamesHeader, idatCsvBirthdayHeader, idatCsvCitizenshipHeader,
-        idatCsvGenderHeader};
+
+    String[] tempIdatHeaders = {identityCsvFirstNameHeader, identityCsvLastNameHeader,
+        identityCsvPreviousNamesHeader, identityCsvBirthdayHeader, identityCsvCitizenshipHeader,
+        identityCsvGenderHeader};
     Arrays.stream(tempIdatHeaders).forEach(header -> {
       if (header != null) {
         this.idatHeaders.add(header);
       }
     });
-    //headerPatientFieldsMap.put(idatCsvFirstNameHeader, ((patientFields, value) -> patientFields.setFirstName(value)));
-    headerPatientFieldsMap.put(idatCsvFirstNameHeader, PatientFields::setFirstName);
-    headerPatientFieldsMap.put(idatCsvLastNameHeader, PatientFields::setLastName);
-    headerPatientFieldsMap.put(idatCsvPreviousNamesHeader, PatientFields::setPreviousNames);
-    headerPatientFieldsMap.put(idatCsvBirthdayHeader, PatientFields::setBirthdate);
-    headerPatientFieldsMap.put(idatCsvCitizenshipHeader, PatientFields::setCitizenship);
-    headerPatientFieldsMap.put(idatCsvGenderHeader, PatientFields::setGender);
+
+    headerPatientFieldsMap.put(identityCsvFirstNameHeader, PatientFields::setFirstName);
+    headerPatientFieldsMap.put(identityCsvLastNameHeader, PatientFields::setLastName);
+    headerPatientFieldsMap.put(identityCsvPreviousNamesHeader, PatientFields::setPreviousNames);
+    headerPatientFieldsMap.put(identityCsvBirthdayHeader, (patientField, value) ->
+        patientField.setBirthdate(patientFieldsUtils.convertBirthdayToIdManager(value)));
+    headerPatientFieldsMap.put(identityCsvCitizenshipHeader, PatientFields::setCitizenship);
+    headerPatientFieldsMap.put(identityCsvGenderHeader, (patientField, value) ->
+        patientField.setGender(patientFieldsUtils.covertGenderToIdManager(value)));
   }
 
   @Override
   public void execute(DelegateExecution delegateExecution) throws Exception {
     logger.info("Generate pseudonyms");
     PathsBundle pathsBundle = PathsBundleUtils.getPathsBundleVariable(delegateExecution);
-    if (containsAlreadyIdManagerPseudonymIdType(pathsBundle)) {
+    if (containsAlreadyIdManagerPseudonymIdType(pathsBundle)) { // Already pseudonymized
       removeIdatAndTempIdHeaders(pathsBundle);
-    } else if (localIdCsvFilename != null) {
+    } else if (identityCsvLocalId != null) { // Local Id
       fetchPseudonymsFromLocalId(pathsBundle);
-    } else if (idatCsvFilename != null) {
+    } else { //IDAT
       fetchPseudonymsFromIdat(pathsBundle);
     }
   }
 
   private void fetchPseudonymsFromLocalId(PathsBundle pathsBundle)
       throws CsvUpdaterFactoryException {
-    CsvUpdater csvUpdater = createCsvUpdater(pathsBundle, localIdCsvFilename);
+    CsvUpdater csvUpdater = createCsvUpdater(pathsBundle, identityCsvFilename);
     //TODO
   }
 
   private void fetchPseudonymsFromIdat(PathsBundle pathsBundle)
       throws CsvUpdaterFactoryException, CsvUpdaterException, CsvReaderException, IOException {
-    CsvUpdater csvUpdater = createCsvUpdater(pathsBundle, idatCsvFilename);
+    CsvUpdater csvUpdater = createCsvUpdater(pathsBundle, identityCsvFilename);
     // Add temporal id as column. This works as pivot.
     csvUpdater.applyConsumer(new TempIdCsvRecordHeaderValuesConsumer());
     // Read IDAT and pseudonomyze.
     Map<Integer, String> patIdPseudonymMap = pseudonymisationClient.fetchPatIdPseudonym(
-        readIdat(pathsBundle, idatCsvFilename));
+        readIdat(pathsBundle, identityCsvFilename));
     // Add Pseudonyms to file.
     csvUpdater.addPivotedCsvRecordHeaderValues(
         createPivotedCsvRecordHeaderValues(patIdPseudonymMap));
@@ -130,27 +135,24 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
   private CsvUpdater createCsvUpdater(PathsBundle pathsBundle, String filename)
       throws CsvUpdaterFactoryException {
     return csvUpdaterFactory.createCsvUpdater(
-        createCsvReaderParametersForUpdater(pathsBundle, idatCsvFilename));
+        createCsvReaderParametersForUpdater(pathsBundle, identityCsvFilename));
   }
 
   private void removeIdatAndTempIdHeaders(PathsBundle pathsBundle)
       throws CsvUpdaterException, CsvUpdaterFactoryException {
-    String filename = getActiveFilename();
-    if (filename != null) {
-      CsvUpdater csvUpdater = createCsvUpdater(pathsBundle, filename);
-      removeIdatAndTempIdHeaders(csvUpdater);
-    }
+    CsvUpdater csvUpdater = createCsvUpdater(pathsBundle, identityCsvFilename);
+    removeIdatAndTempIdHeaders(csvUpdater);
   }
 
   private void removeIdatAndTempIdHeaders(CsvUpdater csvUpdater) throws CsvUpdaterException {
     Set<String> columnsToDelete = (removeIdat) ? getIdatAndTempIdHeaders()
-        : new HashSet<>(Arrays.asList(IDAT_CSV_TEMP_HEADER));
+        : new HashSet<>(Arrays.asList(IDENTITY_CSV_TEMP_HEADER));
     csvUpdater.deleteColumns(columnsToDelete);
   }
 
   private Set<String> getIdatAndTempIdHeaders() {
     List<String> headers = new ArrayList<>();
-    headers.add(IDAT_CSV_TEMP_HEADER);
+    headers.add(IDENTITY_CSV_TEMP_HEADER);
     headers.addAll(idatHeaders);
     return new HashSet<>(headers);
   }
@@ -158,11 +160,11 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
   private PivotedCsvRecordHeaderValues createPivotedCsvRecordHeaderValues(
       Map<Integer, String> patIdPseudonymMap) {
     PivotedCsvRecordHeaderValues pivotedCsvRecordHeaderValues = new PivotedCsvRecordHeaderValues(
-        IDAT_CSV_TEMP_HEADER);
+        IDENTITY_CSV_TEMP_HEADER);
     patIdPseudonymMap.keySet().forEach(patId -> {
       CsvRecordHeaderValues csvRecordHeaderValues = new CsvRecordHeaderValues();
       Map<String, String> headerValueMap = csvRecordHeaderValues.getHeaderValueMap();
-      headerValueMap.put(IDAT_CSV_TEMP_HEADER, patId.toString());
+      headerValueMap.put(IDENTITY_CSV_TEMP_HEADER, patId.toString());
       headerValueMap.put(idManagerPseudonymIdType, patIdPseudonymMap.get(patId));
       pivotedCsvRecordHeaderValues.addCsvRecordHeaderValues(csvRecordHeaderValues);
     });
@@ -175,13 +177,14 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
 
     @Override
     public void accept(CsvRecordHeaderValues csvRecordHeaderValues) throws CsvUpdaterException {
-      csvRecordHeaderValues.getHeaderValueMap().put(IDAT_CSV_TEMP_HEADER, (counter++).toString());
+      csvRecordHeaderValues.getHeaderValueMap()
+          .put(IDENTITY_CSV_TEMP_HEADER, (counter++).toString());
     }
 
     @Override
     public CsvRecordHeaderOrder prepareHeaders(CsvRecordHeaderOrder csvRecordHeaderOrder)
         throws CsvUpdaterException {
-      csvRecordHeaderOrder.addHeaderAtLastPosition(IDAT_CSV_TEMP_HEADER);
+      csvRecordHeaderOrder.addHeaderAtLastPosition(IDENTITY_CSV_TEMP_HEADER);
       return csvRecordHeaderOrder;
     }
   }
@@ -202,7 +205,7 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
         createCsvReaderParametersToReadIdat(pathsBundle, filename))) {
       csvReader.readCsvRecordHeaderValues().forEach(csvRecordHeaderValues -> {
         Patient patient = createPatient(csvRecordHeaderValues);
-        Integer patId = Integer.valueOf(csvRecordHeaderValues.getValue(IDAT_CSV_TEMP_HEADER));
+        Integer patId = Integer.valueOf(csvRecordHeaderValues.getValue(IDENTITY_CSV_TEMP_HEADER));
         patIdPatientMap.put(patId, patient);
       });
     }
@@ -233,18 +236,14 @@ public class PseudonymGeneratorDelegate implements JavaDelegate {
 
   private boolean containsAlreadyIdManagerPseudonymIdType(PathsBundle pathsBundle)
       throws IOException {
-    String filename = getActiveFilename();
-    if (filename != null) {
-      try (Stream<String> lines = Files.lines(pathsBundle.getPath(filename))) {
+    Path path = pathsBundle.getPath(identityCsvFilename);
+    if (Files.exists(path)) {
+      try (Stream<String> lines = Files.lines(path)) {
         Optional<String> first = lines.findFirst();
         return (first.isPresent()) ? first.get().contains(idManagerPseudonymIdType) : false;
       }
     }
     return false;
-  }
-
-  private String getActiveFilename() {
-    return (idatCsvFilename != null) ? idatCsvFilename : localIdCsvFilename;
   }
 
 }